requirements.txt
altair==5.0.1
appdirs==1.4.4
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.5
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
Cython==3.0.0
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.3.2
fonttools==4.40.0
frozendict==2.3.8
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
gspread==5.10.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
multitasking==0.0.11
mysql-connector-python==8.0.33
numpy==1.25.0
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
pandas-datareader==0.10.0
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.5.0
protobuf==3.20.3
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyOpenSSL==23.2.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pyzmq==25.0.0
referencing==0.29.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.8.10
rsa==4.9
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
statsmodels==0.14.0
streamlit==1.24.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.6
xlrd==2.0.1
xlwings==0.30.10
yfinance==0.2.24
zipp==3.16.2
